Fix typo in the name of CapacityBaseRandomPolicy
